Backend/long overflow fix v2 and get Retail Detail

+++ src/main/java/com/code/monks/csm/dto/analytics/projection/SupplierAmountView.java
@@ -1,7 +1,9 @@
 package com.code.monks.csm.dto.analytics.projection;
 
+import java.math.BigInteger;
+
 public interface SupplierAmountView {
 
     Integer getSupplierId();
-    Long getAmount();
+    BigInteger getAmount();
 }
